# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Paperless-LLM API
  version: 1.0.0
  description: |
    REST API for the Paperless-LLM document processing system. 
    
    This API enables automated document processing workflows using Large Language Models (LLMs)
    to generate document metadata such as titles, tags, and summaries. It supports both automated
    and approval-based workflows with comprehensive job tracking, queue management, and step-level
    retry mechanisms.
    
    ## Key Features
    - **Job Submission**: Submit batch processing jobs for multiple documents
    - **Workflow Management**: Track job progress through multi-step workflows
    - **Approval System**: Review and approve/reject LLM-generated suggestions
    - **Queue Monitoring**: Real-time visibility into processing queues
    - **Prompt Customization**: Customize LLM prompts for different processing steps
    - **Health Monitoring**: Monitor system and service health status
    - **Retry Mechanism**: Automatic retry with exponential backoff for failed steps
    
    ## Architecture
    The system follows a clean architecture pattern with:
    - Domain-Driven Design principles
    - Step-based workflow orchestration
    - Transaction management for data consistency
    - RFC 7807 Problem Details for error responses
  contact:
    name: Paperless-LLM Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: API base path
...
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: List jobs with pagination
      description: Retrieve a paginated list of jobs, optionally filtered by job state
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          description: Maximum number of jobs to return (default 50, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor from a previous response
          schema:
            type: string
        - name: state
          in: query
          description: Filter jobs by state
          schema:
            $ref: '#/components/schemas/JobState'
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Jobs
      summary: Submit batch job
      description: |
        Submit one or more documents for processing. Each document can have multiple job types.
        Jobs are processed asynchronously through a multi-step workflow.
      operationId: submitJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchJobRequest'
            examples:
              single-document-single-job:
                summary: Single document, single job type
                value:
                  documents:
                    - documentId: "123"
                      jobTypes: ["automated"]
                      fields: ["title"]
              multiple-documents:
                summary: Multiple documents with different job types
                value:
                  documents:
                    - documentId: "123"
                      jobTypes: ["title"]
                      fields: ["title", "tags"]
                    - documentId: "456"
                      jobTypes: ["title", "tag"]
                      fields: ["title", "tags"]
              with-approval:
                summary: Job requiring approval
                value:
                  documents:
                    - documentId: "789"
                      jobTypes: ["approval"]
      responses:
        '201':
          description: Jobs submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /jobs/stats:
    get:
      tags:
        - Jobs
      summary: Get job statistics
      description: Returns count of jobs grouped by their current state
      operationId: getJobStats
      responses:
        '200':
          description: Job statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /approvals:
    get:
      tags:
        - Approvals
      summary: List pending approvals
      description: Retrieve a paginated list of steps awaiting approval decisions
      operationId: listApprovals
      parameters:
        - name: limit
          in: query
          description: Maximum number of approvals to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor from a previous response
          schema:
            type: string
      responses:
        '200':
          description: List of pending approvals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /approvals/stats:
    get:
      tags:
        - Approvals
      summary: Get approval statistics
      description: Returns count of pending approval requests
      operationId: getApprovalStats
      responses:
        '200':
          description: Approval statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /prompts:
    get:
      tags:
        - Prompts
      summary: List all prompts
      description: Retrieve all LLM prompt templates configured in the system
      operationId: listPrompts
      responses:
        '200':
          description: List of prompts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptsListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /prompts/{stepType}:
    put:
      tags:
        - Prompts
      summary: Update or create prompt
      description: |
        Update an existing prompt template or create a new one for the specified step type.
        The version number is automatically incremented.
      operationId: updatePrompt
      parameters:
        - name: stepType
          in: path
          required: true
          description: Step type for the prompt
          schema:
            $ref: '#/components/schemas/StepType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                template:
                  type: string
                  description: Prompt template text with variable placeholders
                  example: "Generate a concise title for this document:\n\n{{documentContent}}"
              required:
                - template
      responses:
        '200':
          description: Prompt updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /queue/stats:
    get:
      tags:
        - Queue
      summary: Get unified queue statistics
      description: Returns aggregated statistics for all queue items across the system
      operationId: getQueueStats
      responses:
        '200':
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /documents:
    get:
      tags:
        - Documents
      summary: List documents by tag
      description: Retrieve documents from Paperless-NGX filtered by tag ID
      operationId: listDocuments
      parameters:
        - name: limit
          in: query
          description: Maximum number of documents to return (upper limit)
          schema:
            type: number
        - name: cursor
          in: query
          description: Cursor returned from prior queries
          schema:
            type: string
        - name: tag
          in: query
          required: true
          description: Tag ID to filter documents
          schema:
            type: string
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                required:
                  - documents
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    WorkflowType:
      type: string
      enum:
        - approval
        - automated
      description: Type of workflow processing mode
    JobState:
      type: string
      enum:
        - pending
        - llm_processing
        - pending_approval
        - updating_document
        - completed
        - failed
        - rejected
      description: Current state of a job in its workflow
    StepType:
      type: string
      enum:
        - LLM_GENERATE_TITLE
        - LLM_GENERATE_FIELDS
        - LLM_GENERATE_TAGS
        - LLM_GENERATE_CORRESPONDENT
        - LLM_GENERATE_DOCUMENT_TYPE
        - LLM_GENERATE_CREATED_DATE
        - REQUIRE_APPROVAL
        - UPDATE_DOCUMENT
        - REMOVE_TAGS
      description: Type of workflow step
...
    BatchJobRequest:
      type: object
      properties:
        documents:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                description: ID of the document in Paperless-NGX
                example: "123"
              jobType:
                $ref: '#/components/schemas/WorkflowType'
              fields:
                  $ref: '#/components/schemas/LLMGenerateableFields'
            required:
              - documentId
              - jobType
              - fields
          minItems: 1
      required:
        - documents
    LLMGenerateableFields:
      type: array
      items:
        type: string
        enum: ["title", "tags", "correspondent", "document_type", "created_date"]


    # Response Schemas
    JobSubmissionResponse:
      type: object
      properties:
        submitted:
          type: integer
          description: Number of jobs successfully submitted
          example: 2
        jobs:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                example: "123"
              jobType:
                type: string
                example: title
              id:
                type: string
                format: uuid
                example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            required:
              - documentId
              - jobType
              - id
      required:
        - submitted
        - jobs
    JobResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        fields:
          $ref: '#/components/schemas/LLMGenerateableFields'
        documentId:
          type: string
          example: "123"
        jobType:
          $ref: '#/components/schemas/WorkflowType'
        status:
          $ref: '#/components/schemas/JobState'
        errorMessage:
          type: string
          nullable: true
          description: Error message if job failed
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-11T10:15:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-11T10:30:00Z"
        documentActions:
          type: array
          items:
            $ref: '#/components/schemas/DocumentAction'
      required:
        - id
        - documentId
        - jobType
        - status
        - createdAt
        - updatedAt
        - documentActions
    DocumentAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actionType:
          type: string
          description: Type of document action (e.g., 'update_title', 'add_tag')
          example: update_title
        oldValue:
          type: string
          nullable: true
          description: Previous value before the action
          example: "Untitled Document"
        newValue:
          type: string
          nullable: true
          description: New value after the action
          example: "Invoice 2026-05-11"
      required:
        - id
        - actionType
        - oldValue
        - newValue
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobResponse'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page of results
      required:
        - jobs
        - nextCursor
...
    JobStats:
      type: object
      properties:
        pending:
          type: integer
          minimum: 0
          example: 5
        llmProcessing:
          type: integer
          minimum: 0
          example: 3
        pendingApproval:
          type: integer
          minimum: 0
          example: 2
        updatingDocument:
          type: integer
          minimum: 0
          example: 1
        completed:
          type: integer
          minimum: 0
          example: 42
        failed:
          type: integer
          minimum: 0
          example: 1
        rejected:
          type: integer
          minimum: 0
          example: 0
      required:
        - pending
        - llmProcessing
        - pendingApproval
        - updatingDocument
        - completed
        - failed
        - rejected
    ApprovalsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalItem'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page of results
      required:
        - items
        - nextCursor
    ApprovalItem:
      type: object
      properties:
        stepId:
          type: string
          format: uuid
          example: "c3d4e5f6-a7b8-9012-cdef-123456789012"
        jobId:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        documentId:
          type: string
          example: "123"
        paperlessUrl:
          type: string
          format: uri
          description: URL to view the document in Paperless-NGX
          example: "http://paperless.local/documents/123"
        jobType:
          type: string
          example: title
        proposedActions:
          type: array
          items:
            type: object
            properties:
              actionType:
                type: string
                example: update_title
              oldValue:
                type: string
                example: "Untitled"
              newValue:
                type: string
                example: "Invoice 2026-05-11"
            required:
              - actionType
              - oldValue
              - newValue
        possibleDecisions:
          type: array
          items:
            type: string
          example: ["approve", "reject"]
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:15:00Z"
      required:
        - stepId
        - jobId
        - documentId
        - paperlessUrl
        - jobType
        - proposedActions
        - possibleDecisions
        - createdAt
    ApprovalStats:
      type: object
      properties:
        pendingCount:
          type: integer
          minimum: 0
          example: 3
      required:
        - pendingCount
    PromptsListResponse:
      type: object
      properties:
        prompts:
          type: array
          items:
            $ref: '#/components/schemas/PromptResponse'
      required:
        - prompts
    PromptResponse:
      type: object
      properties:
        stepType:
          $ref: '#/components/schemas/StepType'
        template:
          type: string
          description: Prompt template with variable placeholders
          example: "Generate a concise title for this document:\n\n{{documentContent}}"
        version:
          type: integer
          minimum: 1
          description: Version number of the prompt template
          example: 1
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-11T09:00:00Z"
      required:
        - stepType
        - template
        - version
        - updatedAt
    QueueStats:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
          example: 15
        pending:
          type: integer
          minimum: 0
          example: 10
        processing:
          type: integer
          minimum: 0
          example: 2
        completed:
          type: integer
          minimum: 0
          example: 2
        failed:
          type: integer
          minimum: 0
          example: 1
      required:
        - total
        - pending
        - processing
        - completed
        - failed
...
    Document:
      type: object
      properties:
        id:
          type: string
          example: "123"
        title:
          type: string
          example: "Invoice 2026-05-11"
        content:
          type: string
          description: Full text content of the document
      required:
        - id
        - title
        - content
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "about:blank"
        title:
          type: string
          description: Short, human-readable summary of the problem
          example: "Bad Request"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "Validation failed for field 'documentId'"
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: "/jobs/abc123"
      required:
        - type
        - title
        - status
  responses:
    BadRequest:
      description: Bad request - validation error or malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "about:blank"
            title: "Bad Request"
            status: 400
            detail: "Validation failed: documentId is required"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "about:blank"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing the request"